Fix small bug
