+++ crates/pop-cli/Cargo.toml
@@ -21,7 +21,6 @@ reqwest.workspace = true
 serde_json.workspace = true
 tempfile.workspace = true
 tokio.workspace = true
-toml_edit.workspace = true
 url.workspace = true
 
 # pop-cli
