+++ crates/pop-cli/src/commands/new/pallet.rs
@@ -1,13 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0
 
-use crate::{
-	common::helpers::{add_crate_to_workspace, find_workspace_toml},
-	style::Theme,
-};
+use crate::style::Theme;
 use clap::Args;
 use cliclack::{clear_screen, confirm, intro, outro, outro_cancel, set_theme};
 use console::style;
 use pop_parachains::{create_pallet_template, resolve_pallet_path, TemplatePalletConfig};
+use pop_common::manifest::{add_crate_to_workspace, find_workspace_toml};
 use std::fs;
 
 #[derive(Args)]
