+++ crates/pop-common/Cargo.toml
@@ -23,6 +23,7 @@ tar.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+toml_edit.workspace = true
 url.workspace = true
 
 [dev-dependencies]
